Use URL helper classes from jpi/utils
